Changed to docker setup and converting to cryptodome for aes/hashing
