+++ api/tacticalrmm/requirements.txt
@@ -20,7 +20,7 @@ more-itertools==8.4.0
 packaging==20.4
 psycopg2-binary==2.8.5
 pycparser==2.20
-pycryptodome==3.9.7
+pycryptodome==3.9.8
 pyotp==2.3.0
 pyparsing==2.4.7
 pytz==2020.1
