+++ docker/docker-compose.dev.yml
@@ -40,6 +40,7 @@ services:
     depends_on: 
       - db
       - venv
+      - meshcentral
   
   # Container for Celery worker service
   celery-service:
