+++ docker/meshcentral/entry.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 node node_modules/meshcentral --createaccount ${MESH_USER} --pass ${MESH_PASS} --email ${EMAIL_USER}
 node node_modules/meshcentral --adminaccount ${MESH_USER}
